requirements.txt
statsmodels
pmdarima
pandas
requests
anthropic
python-dotenv
